Allow setting PKCS11_MODULE path via environment variable
If I read the code right, libdigidocpp and/or qdigidoc4 take
`PKCS11_MODULE` as file name and attempt to load it based on Qt's
`applicationDirPath()`.

Either way, on platforms such as OpenBSD, the current code fails to
find a suitable module and thus `digidoc-tool` and `qdigidoc4` are
unable to interact with smart cards.

Copying what `modules/VersionInfo.cmake` does, allow passing an
environment variable `PKCS11_MODULE_PATH` such that package maintainers
(like me) may pass an absoloute path to the required PKCS11 module at
build-time.

This seems suitable as environment variables can be set easily and do
not need further handling in CMakeLists.txt, as would be the case with
a proper cmake variable.

Signed-off-by: Klemens Nanni <klemens@posteo.de>